+++ src/Louis/Fluency/FluentExtensions.cs
@@ -27,7 +27,7 @@ public static class FluentExtensions
     /// <returns>A reference to <paramref name="this"/> after <paramref name="action"/> returns.</returns>
     public static T Invoke<T>(this T @this, Action<T> action)
     {
-        Arg.NotNull(action).Value(@this);
+        Require.NotNull(action).Value(@this);
         return @this;
     }
 
@@ -40,7 +40,7 @@ public static class FluentExtensions
     /// <param name="then">The action to perform on <paramref name="this"/> if <paramref name="condition"/> is <see langword="true"/>.</param>
     /// <returns>A reference to <paramref name="this"/> after <paramref name="then"/>, if called, returns.</returns>
     public static T If<T>(this T @this, bool condition, FluentAction<T> then)
-        => condition ? Arg.NotNull(then).Value(@this) : @this;
+        => condition ? Require.NotNull(then).Value(@this) : @this;
 
     /// <summary>
     /// Invokes one of two actions on an object according to a boolean condition, then returns the same object.
@@ -52,7 +52,7 @@ public static class FluentExtensions
     /// <param name="else">The action to perform on <paramref name="this"/> if <paramref name="condition"/> is <see langword="false"/>.</param>
     /// <returns>A reference to <paramref name="this"/> after either <paramref name="then"/> or <paramref name="else"/> returns.</returns>
     public static T IfElse<T>(this T @this, bool condition, FluentAction<T> then, FluentAction<T> @else)
-        => condition ? Arg.NotNull(then).Value(@this) : Arg.NotNull(@else).Value(@this);
+        => condition ? Require.NotNull(then).Value(@this) : Require.NotNull(@else).Value(@this);
 
     /// <summary>
     /// Selects an action to invoke on an object by comparing a given value against a list of comparands, then returns the object.
@@ -98,7 +98,7 @@ public static class FluentExtensions
     public static T Switch<T, TValue>(this T @this, TValue value, FluentAction<T>? @default, params (TValue Comparand, FluentAction<T>? Action)[] cases)
         where TValue : IEquatable<TValue>
     {
-        foreach (var (comparand, action) in Arg.NotNull(cases).Value)
+        foreach (var (comparand, action) in Require.NotNull(cases).Value)
         {
             if (value.Equals(comparand))
             {
@@ -123,8 +123,8 @@ public static class FluentExtensions
     /// <seealso cref="ForEach{T,TElement}(T,ReadOnlySpan{TElement},FluentAction{T,TElement})"/>
     public static T ForEach<T, TElement>(this T @this, IEnumerable<TElement> sequence, FluentAction<T, TElement> action)
     {
-        _ = Arg.NotNull(action);
-        foreach (var item in Arg.NotNull(sequence).Value)
+        _ = Require.NotNull(action);
+        foreach (var item in Require.NotNull(sequence).Value)
         {
             @this = action(@this, item);
         }
@@ -151,9 +151,9 @@ public static class FluentExtensions
     /// <seealso cref="ForEach{T,TElement}(T,ReadOnlySpan{TElement},FluentAction{T,TElement,int})"/>
     public static T ForEach<T, TElement>(this T @this, IEnumerable<TElement> sequence, FluentAction<T, TElement, int> action)
     {
-        _ = Arg.NotNull(action);
+        _ = Require.NotNull(action);
         var index = 0;
-        foreach (var item in Arg.NotNull(sequence).Value)
+        foreach (var item in Require.NotNull(sequence).Value)
         {
             @this = action(@this, item, index++);
         }
@@ -175,7 +175,7 @@ public static class FluentExtensions
     /// <seealso cref="ForEach{T,TElement}(T,IEnumerable{TElement},FluentAction{T,TElement})"/>
     public static T ForEach<T, TElement>(this T @this, ReadOnlySpan<TElement> span, FluentAction<T, TElement> action)
     {
-        _ = Arg.NotNull(action);
+        _ = Require.NotNull(action);
         foreach (var item in span)
         {
             @this = action(@this, item);
@@ -202,7 +202,7 @@ public static class FluentExtensions
     /// <seealso cref="ForEach{T,TElement}(T,IEnumerable{TElement},FluentAction{T,TElement,int})"/>
     public static T ForEach<T, TElement>(this T @this, ReadOnlySpan<TElement> span, FluentAction<T, TElement, int> action)
     {
-        _ = Arg.NotNull(action);
+        _ = Require.NotNull(action);
         for (var i = 0; i < span.Length; i++)
         {
             @this = action(@this, span[i], i);