+++ src/Louis/Logging/LoggerExtensions-Log.cs
@@ -26,7 +26,7 @@ partial class LoggerExtensions
     /// <param name="message">The log message to write.</param>
     public static void Log(this ILogger @this, LogLevel logLevel, string message)
     {
-        _ = Arg.NotNull(@this);
+        _ = Require.NotNull(@this);
         if (@this.IsEnabled(logLevel))
         {
             @this.Log(logLevel, NullEventId, message, null, StaticMessageFormatter);
@@ -42,7 +42,7 @@ partial class LoggerExtensions
     /// <param name="message">The log message to write.</param>
     public static void Log(this ILogger @this, LogLevel logLevel, EventId eventId, string message)
     {
-        _ = Arg.NotNull(@this);
+        _ = Require.NotNull(@this);
         if (@this.IsEnabled(logLevel))
         {
             @this.Log(logLevel, eventId, message, null, StaticMessageFormatter);
@@ -58,7 +58,7 @@ partial class LoggerExtensions
     /// <param name="message">The log message to write.</param>
     public static void Log(this ILogger @this, LogLevel logLevel, Exception? exception, string message)
     {
-        _ = Arg.NotNull(@this);
+        _ = Require.NotNull(@this);
         if (@this.IsEnabled(logLevel))
         {
             @this.Log(logLevel, NullEventId, message, exception, StaticMessageFormatter);
@@ -75,7 +75,7 @@ partial class LoggerExtensions
     /// <param name="message">The log message to write.</param>
     public static void Log(this ILogger @this, LogLevel logLevel, EventId eventId, Exception? exception, string message)
     {
-        _ = Arg.NotNull(@this);
+        _ = Require.NotNull(@this);
         if (@this.IsEnabled(logLevel))
         {
             @this.Log(logLevel, eventId, message, exception, StaticMessageFormatter);